work around ReactiveX/RxSwift#2125
https://github.com/ReactiveX/RxSwift/issues/2125

macOS prior to 10.12 does not support weak references to NSTextViews,
and Xcode 11.4 beta enforces this as a compile-time error if the build
target's deployment target is earlier than 10.12. RxCocoa 5.0.1's
deployment target is 10.10.

To work around the problem, I've marked this class, and the methods that
use it, as only available on 10.12 or later. However, that's not enough
to convince the compiler! So I've also changed the type of the weak
reference to AnyObject?, and hidden the AnyObject? reference behind a
computed property.